requirements.txt
# Core ML libraries
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0

# Web framework
flask==2.3.3
werkzeug==2.3.7

# Data science (for training/analysis)
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tqdm==4.66.1

# Image processing
opencv-python==4.8.0.76

# Development tools (optional)
jupyter==1.0.0
notebook==6.5.4
